Split polyfills out to separate file
